review fixes move tap detection logic to skiko remove redunant code remove duplicate
