+++ compose/foundation/foundation/src/commonMain/kotlin/androidx/compose/foundation/text/selection/TextSelectionMouseDetector.kt
@@ -21,6 +21,7 @@ import androidx.compose.foundation.gestures.drag
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.input.pointer.*
 import androidx.compose.ui.platform.ViewConfiguration
+import androidx.compose.ui.unit.dp
 import androidx.compose.ui.util.fastAll
 
 // * Without shift it starts the new selection from the scratch.
@@ -40,26 +41,24 @@ internal interface MouseSelectionObserver {
     fun onDrag(dragPosition: Offset, adjustment: SelectionAdjustment): Boolean
 }
 
-// Distance in pixels between consecutive click positions to be considered them as clicks sequence
-internal const val ClicksSlop = 100.0
-
-private class ClicksCounter(
-    private val viewConfiguration: ViewConfiguration
+internal class ClicksCounter(
+    private val viewConfiguration: ViewConfiguration,
+    private val clicksSlop: Float // Distance in pixels between consecutive click positions to be considered them as clicks sequence
 ) {
     var clicks = 0
     var prevClick: PointerInputChange? = null
-    fun update(event: PointerEvent) {
-        val currentPrevClick = prevClick
-        val newClick = event.changes[0]
-        if (currentPrevClick != null &&
-            timeIsTolerable(currentPrevClick, newClick) &&
-            positionIsTolerable(currentPrevClick, newClick)
+    fun update(event: PointerInputChange) {
+        val currentPrevEvent = prevClick
+        // Here and further event means upcoming event (new)
+        if (currentPrevEvent != null &&
+            timeIsTolerable(currentPrevEvent, event) &&
+            positionIsTolerable(currentPrevEvent, event)
         ) {
             clicks += 1
         } else {
             clicks = 1
         }
-        prevClick = newClick
+        prevClick = event
     }
 
     fun timeIsTolerable(prevClick: PointerInputChange, newClick: PointerInputChange): Boolean {
@@ -69,7 +68,7 @@ private class ClicksCounter(
 
     fun positionIsTolerable(prevClick: PointerInputChange, newClick: PointerInputChange): Boolean {
         val diff = newClick.position - prevClick.position
-        return diff.getDistance() < ClicksSlop
+        return diff.getDistance() < clicksSlop
     }
 }
 
@@ -77,11 +76,11 @@ internal suspend fun PointerInputScope.mouseSelectionDetector(
     observer: MouseSelectionObserver
 ) {
     awaitEachGesture {
-        val clicksCounter = ClicksCounter(viewConfiguration)
+        val clicksCounter = ClicksCounter(viewConfiguration, clicksSlop = 50.dp.toPx())
         while (true) {
             val down = awaitMouseEventDown()
-            clicksCounter.update(down)
             val downChange = down.changes[0]
+            clicksCounter.update(downChange)
             if (down.keyboardModifiers.isShiftPressed) {
                 val started = observer.onExtend(downChange.position)
                 if (started) {
